Attempt to fix test by not including two classes in same file
